changed linkedlist to a new file
