+++ linked lists/node.py
@@ -4,36 +4,7 @@ class Node:
         self.data = data    # can be any type
         self.next = None    # type : Node
         
-class LinkedList:
-    def __init__(self, head):
-        self.head = head    # type : Node
-    
-    # inserts a the new data at the begining of the list
-    def insertAtBegin(self, data = 'd' ):
-        n1 = Node(data)
-        n1.next = self.head
-        self.head = n1
-        
-        
-        
-    def insertAtIndex(self, data, index):
-        pass
-    
-    # ex 1 
-    # insert an element 1 place after the end.
-    # example-    lst: 1 -> 3 -> 9 
-    #             lst.insertOneAfterHead(10)
-    #             lst: 1 -> 10 -> 3 -> 9 
-    def insertOneAfterHead(self, data): # @TODO
-        pass
-    
-    # ex 2: 
-    # insert an at the end of the list
-    # example-    lst: 1 -> 3 -> 9 
-    #             lst.insertLast(10)
-    #             lst: 1 -> 3 -> 9 -> 10
-    def insertLast(self, data): # @TODO
-        pass
+
 
 if __name__ == '__main__':
     n1 = Node(10)
